+++ src/main/java/com/vitalisplus/vitalixplus_pedido_service/infrastructure/dto/request/PedidoRequestDTO.java
@@ -15,6 +15,6 @@ public class PedidoRequestDTO {
     private String direccionEntrega;
     private Double costoEnvio;
     private Double costoPedido;
-    private List<Integer> listaDeProductos;
+    private List<Long> listaDeProductos;
     private Double totalPagar;
 }
