+++ src/main/java/com/vitalisplus/vitalixplus_pedido_service/infrastructure/dto/response/PedidoResponseDTO.java
@@ -21,6 +21,6 @@ public class PedidoResponseDTO {
     private String direccionEntrega;
     private Double costoEnvio;
     private Double costoPedido;
-    private List<Integer> listaDeProductos;
+    private List<Long> listaDeProductos;
     private Double totalPagar;
 }
